requirements.txt
# developed on python 3.9

category_encoders==2.2.2
jinja2
pyod==0.9.3
shap==0.40.0
matplotlib==3.4.3
dask-ml[complete]==1.9.0
scikit-learn==0.24.2
scikit-learn-extra==0.2.0
plotly==5.3.1
pandas==1.3.3
numpy==1.19.5
lime==0.2.0.1
# torch==1.12.0
torchvision==0.13.1
captum==0.5.0
scipy==1.7.1
saliency==0.2.0
torchinfo
openpyxl==3.0.7
opencv-python==4.5.5.62
datasets[vision]
wandb==0.13.9
pytorch-lightning==1.6.5
torchmetrics==0.9.3
notebook==6.5.4

# install torch gpu
# start venv\Scripts\activate  |  source venv/bin/activate
# pip install torch==1.12.0+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html

# package location
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html